+++ pytool/__init__.py
@@ -14,12 +14,13 @@
     limitations under the License.
 
 """
-__version__ = '2.1.0'
+__version__ = '2.2.0-dev'
 
 
 from pytool import (
         cmd,
         json,
         lang,
+        proxy,
         time,
         )